pyproject.toml
[tool.poetry]
name = "qna-rewind"
version = "0.1.0"
description = ""
authors = ["jihun.Lee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.0"
django-filter = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
